Organize built-in middleware by putting them into packages
