+++ slack_bolt/middleware/ignoring_self_events/__init__.py
@@ -1,0 +1,1 @@
+from .ignoring_self_events import IgnoringSelfEvents  # noqa
