+++ slack_bolt/middleware/ignoring_self_events/ignoring_self_events.py
@@ -5,7 +5,7 @@ from slack_bolt.auth import AuthorizationResult
 from slack_bolt.logger import get_bolt_logger
 from slack_bolt.request import BoltRequest
 from slack_bolt.response import BoltResponse
-from .middleware import Middleware
+from slack_bolt.middleware.middleware import Middleware
 
 
 class IgnoringSelfEvents(Middleware):
