+++ slack_bolt/middleware/message_listener_matches/message_listener_matches.py
@@ -1,28 +1,24 @@
 import re
-from typing import Callable, Awaitable, Union, Pattern
+from typing import Callable, Pattern, Union
 
-from slack_bolt.request.async_request import AsyncBoltRequest
+from slack_bolt.request import BoltRequest
 from slack_bolt.response import BoltResponse
-from .async_middleware import AsyncMiddleware
+from slack_bolt.middleware.middleware import Middleware
 
 
-class AsyncMessageListenerMatches(AsyncMiddleware):
+class MessageListenerMatches(Middleware):  # type: ignore
     def __init__(self, keyword: Union[str, Pattern]):
         self.keyword = keyword
 
-    async def async_process(
-        self,
-        *,
-        req: AsyncBoltRequest,
-        resp: BoltResponse,
-        next: Callable[[], Awaitable[BoltResponse]],
+    def process(
+        self, *, req: BoltRequest, resp: BoltResponse, next: Callable[[], BoltResponse],
     ) -> BoltResponse:
         text = req.payload.get("event", {}).get("text", "")
         if text:
             m = re.search(self.keyword, text)
             if m is not None:
                 req.context["matches"] = m.groups()  # tuple
-                return await next()
+                return next()
 
         # As the text doesn't match, skip running the listener
         return resp
