+++ slack_bolt/middleware/ssl_check/__init__.py
@@ -1,0 +1,1 @@
+from .ssl_check import SslCheck  # noqa
